FIX: Background nodes now don't cover paths
After reading this issue, https://github.com/plotly/plotly.py/issues/2345#issuecomment-858396014

We needed to convert the traces, which were Scatter() objects, into Scattergl() objects as they have different drawing priorities.

In the future, we should change the cache to be of ScatterGL objects instead of Scatter. This is a temporary fix